% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bslr.R
\name{BSLR}
\alias{BSLR}
\title{Solve Bayesian simple linear regression.}
\usage{
BSLR(x, y, sigma2, sigma0_2)
}
\arguments{
\item{x}{explanatory variable, n*1 vector.}

\item{y}{response variable, n*1 vector.}

\item{sigma2}{variance of noise, scalar.}

\item{sigma0_2}{variance of the coefficient, scalar.}
}
\value{
The posterior mean and variance, and bayes factor.
}
\description{
Solve Bayesian simple linear regression.
}
