% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredibleSets.R
\name{CredibleSet}
\alias{CredibleSet}
\title{Compute credible sets.}
\usage{
CredibleSet(alpha_mat, rho)
}
\arguments{
\item{alpha_mat}{matrix containing posterior means of gammas.}

\item{rho}{minimal probability that credible sets should reach.}
}
\value{
The required credible sets.
}
\description{
Compute credible sets.
}
