% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBSS.R
\name{IBSS}
\alias{IBSS}
\title{Iterative Bayesian Stepwise Selection.}
\usage{
IBSS(X, Y, sigma2, sigma0_2, L, iter.max = 500)
}
\arguments{
\item{X}{explanatory variable, n*p matrix.}

\item{Y}{response variable, n*1 matrix.}

\item{sigma2}{variance of the noise, scalar.}

\item{sigma0_2}{variance of the coefficient, scalar.}

\item{L}{number of effect variables}

\item{iter.max}{maximum number of iterations}
}
\value{
The estimated posterior parameters.
}
\description{
Iterative Bayesian Stepwise Selection.
}
