% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIP.R
\name{PIP}
\alias{PIP}
\title{Compute posterior inclusion probabilities.}
\usage{
PIP(alpha_mat)
}
\arguments{
\item{alpha_mat}{matrix containing posterior means of gammas.}
}
\value{
Posterior inclusion probabilities.
}
\description{
Compute posterior inclusion probabilities.
}
