% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredibleSets.R
\name{purify_cs}
\alias{purify_cs}
\title{Get purified credible sets}
\usage{
purify_cs(X, cs, purity = 0.5, abs = FALSE)
}
\arguments{
\item{X}{explanatory variable, n*p matrix.}

\item{cs}{credible sets}

\item{purity}{minimal purity as defined in the paper.}

\item{abs}{logical value, indicating whether the criterion should be abs of correlation.}
}
\value{
Purified credible sets.
}
\description{
Get purified credible sets
}
