% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ser.R
\name{SER}
\alias{SER}
\title{Solve single effect regression.}
\usage{
SER(X, Y, sigma2, sigma0_2)
}
\arguments{
\item{X}{explanatory variable, n*p matrix.}

\item{Y}{response variable, n*1 matrix.}

\item{sigma2}{variance of noise, scalar.}

\item{sigma0_2}{variance of the coefficient, scalar.}
}
\value{
The posterior parameters.
}
\description{
Solve single effect regression.
}
